+++ app/javascript/gobierto_dashboards/webapp/components/WidgetIndicatorIndividual.vue
@@ -31,7 +31,7 @@
 </template>
 
 <script>
-import { date, percent } from "lib/shared"
+import { date, percent } from "lib/vue/filters"
 
 export default {
   name: "WidgetIndicatorIndividual",
