+++ app/javascript/gobierto_dashboards/webapp/components/WidgetIndicatorTable.vue
@@ -48,7 +48,7 @@
 </template>
 
 <script>
-import { date, percent } from "lib/shared";
+import { date, percent } from "lib/vue/filters";
 
 const formatValues = (arr, prop, transform = x => x) =>
   arr.reduce(
