+++ app/javascript/gobierto_plans/webapp/components/CustomFieldParagraph.vue
@@ -18,7 +18,7 @@
 </template>
 
 <script>
-import { translate } from "lib/shared";
+import { translate } from "lib/vue/filters";
 
 export default {
   name: "CustomFieldParagraph",
