+++ app/javascript/gobierto_plans/webapp/components/CustomFieldSelection.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script>
-import { translate } from "lib/shared";
+import { translate } from "lib/vue/filters";
 
 export default {
   name: "CustomFieldSelection",
