+++ app/javascript/gobierto_plans/webapp/components/GroupsByTermTableRow.vue
@@ -26,7 +26,7 @@
 </template>
 
 <script>
-import { percent } from "lib/shared";
+import { percent } from "lib/vue/filters";
 import { routes } from "../lib/router";
 
 export default {
