+++ app/javascript/gobierto_plans/webapp/components/Header.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-import { percent } from "lib/shared"
+import { percent } from "lib/vue/filters"
 
 export default {
   name: "Header",
