+++ app/javascript/gobierto_plans/webapp/components/ProjectNativeFields.vue
@@ -36,7 +36,7 @@
 </template>
 
 <script>
-import { date, percent } from "lib/shared";
+import { date, percent } from "lib/vue/filters";
 import { PlansStore } from "../lib/store";
 
 export default {
