+++ app/javascript/gobierto_plans/webapp/lib/mixins/names.js
@@ -1,4 +1,4 @@
-import { translate } from "lib/shared";
+import { translate } from "lib/vue/filters";
 import { PlansStore } from "../store";
 
 export const NamesMixin = {
