+++ app/javascript/gobierto_visualizations/webapp/containers/contracts/AssigneesShow.vue
@@ -24,7 +24,7 @@
 
 <script>
 
-import { VueFiltersMixin } from "lib/shared"
+import { VueFiltersMixin } from "lib/vue/filters"
 
 import Table from "../../components/Table.vue";
 import { EventBus } from "../../mixins/event_bus";
