+++ app/javascript/gobierto_visualizations/webapp/containers/costs/table/TableFirstLevel.vue
@@ -60,7 +60,8 @@
 </template>
 <script>
 import TableHeader from './TableHeader.vue'
-import { VueFiltersMixin } from "lib/shared"
+import { VueFiltersMixin } from "lib/vue/filters"
+
 export default {
   name: "TableFirstLevel",
   components: {
