+++ app/javascript/gobierto_visualizations/webapp/containers/costs/table/TableItem.vue
@@ -346,7 +346,8 @@
 <script>
 import TableHeader from './TableHeader.vue'
 import TableRow from './TableRow.vue'
-import { VueFiltersMixin } from "lib/shared"
+import { VueFiltersMixin } from "lib/vue/filters"
+
 export default {
   name: "TableItem",
   components: {
