+++ app/javascript/gobierto_visualizations/webapp/containers/subsidies/SubsidiesShow.vue
@@ -45,7 +45,7 @@
 
 <script>
 
-import { VueFiltersMixin } from "lib/shared"
+import { VueFiltersMixin } from "lib/vue/filters"
 
 export default {
   name: 'SubsidiesShow',
