+++ app/javascript/lib/visualizations/modules/group_pct_distribution_bars.js
@@ -2,7 +2,7 @@ import { axisTop } from "d3-axis";
 import { scaleThreshold } from "d3-scale";
 import { select } from "d3-selection";
 import { rowChart } from "dc";
-import { truncate } from "lib/shared";
+import { truncate } from "lib/vue/filters";
 
 const d3 = { scaleThreshold, axisTop, select };
 const dc = { rowChart };
